Assistant checkpoint: Fix TypeScript buffer type errors
Assistant generated file changes:
- apps/api/src/index.ts: Fix TypeScript errors by properly typing buffer operations

---

User prompt:

/home/runner/workspace/apps/api/node_modules/ts-node/src/index.ts:859
[api]     return new TSError(diagnosticText, diagnosticCodes, diagnostics);
[api]            ^
[api] TSError: ⨯ Unable to compile TypeScript:
[api] src/index.ts(168,82): error TS2345: Argument of type '16' is not assignable to parameter of type 'BufferEncoding | undefined'.
[api] src/index.ts(172,80): error TS2345: Argument of type '16' is not assignable to parameter of type 'BufferEncoding | undefined'.
[api] src/index.ts(185,32): error TS2769: No overload matches this call.
[api]   The last overload gave the following error.
[api]     Argument of type '[Buffer<ArrayBufferLike>]' is not assignable to parameter of type 'WithImplicitCoercion<string | ArrayLike<number>>'.
[api]       Type '[Buffer<ArrayBufferLike>]' is not assignable to type 'ArrayLike<number>'.
[api]         'number' index signatures are incompatible.
[api]           Type 'Buffer<ArrayBufferLike>' is not assignable to type 'number'.

+++ apps/api/src/index.ts
@@ -165,11 +165,11 @@ app.get("/api/images/storage/:replId/:filename", async (req, res) => {
 
     // Log first few bytes for verification
     if (bytesValue && bytesValue.length > 0) {
-      const firstBytes = Array.from(bytesValue.slice(0, 16)).map(b => b.toString(16).padStart(2, '0')).join(' ');
+      const firstBytes = Array.from(bytesValue.slice(0, 16) as Uint8Array).map(b => b.toString(16).padStart(2, '0')).join(' ');
       console.log(`   - First 16 bytes (hex): ${firstBytes}`);
 
       // Check for common image file signatures
-      const signature = Array.from(bytesValue.slice(0, 4)).map(b => b.toString(16).padStart(2, '0')).join('');
+      const signature = Array.from(bytesValue.slice(0, 4) as Uint8Array).map(b => b.toString(16).padStart(2, '0')).join('');
       console.log(`   - File signature: ${signature}`);
 
       if (signature.startsWith('ffd8')) {
